+++ react-app/src/features/dashboard/index.tsx
@@ -42,7 +42,7 @@ const loader = (queryClient: QueryClient, loginFlag?: boolean) => {
 export const dashboardLoader = loader;
 
 export const Dashboard = () => {
-  const { data: userObj, isLoading } = useGetUser();
+  const { data: oneMacUser, isLoading: isOneMacUserLoading } = useGetUser();
   const { data: userDetails, isLoading: isUserDetailsLoading } = useGetUserDetails();
   const osData = useOsData();
 
@@ -53,76 +53,76 @@ export const Dashboard = () => {
 
   const isAbleToAccessDashboard = () => {
     return (
-      (userObj.user["custom:cms-roles"] || userObj.user["custom:ismemberof"]) &&
-      Object.values(UserRoles).some((role) => userObj.user.role === role)
+      (oneMacUser.user["custom:cms-roles"] || oneMacUser.user["custom:ismemberof"]) &&
+      Object.values(UserRoles).some((role) => oneMacUser.user.role === role)
     );
   };
 
-  if (isLoading || isUserDetailsLoading || osData.tabLoading) {
+  if (isOneMacUserLoading || isUserDetailsLoading) {
     return <LoadingSpinner />;
   }
 
-  if (!userObj?.user || !isAbleToAccessDashboard()) {
+  if (!oneMacUser?.user || !isAbleToAccessDashboard()) {
     return <Navigate to="/" />;
   }
 
   return (
-    <OsProvider
-      value={{
-        data: osData.data,
-        error: osData.error,
-        isLoading: osData.isLoading,
-      }}
-    >
-      <div>
+    <>
+      <div className="flex flex-col w-full self-center mx-auto max-w-screen-xl xs:flex-row justify-between p-4 lg:px-8">
+        <h1 className="text-xl font-bold mb-4 md:mb-0">Dashboard</h1>
+        {isStateUser({ ...oneMacUser.user, role: userDetails.role }) && (
+          <Link
+            to="/new-submission"
+            className="flex items-center text-white font-bold bg-primary border-none px-10 py-2 rounded cursor-pointer"
+          >
+            <span data-testid="new-sub-button" className="mr-2">
+              New Submission
+            </span>
+            <PlusIcon className="w-4 h-4" />
+          </Link>
+        )}
+      </div>
+
+      <Tabs
+        value={osData.state.tab}
+        onValueChange={(tab) => {
+          setLocalStorageCol((prev) => ({
+            ...prev,
+            [osData.state.tab]: osData.state,
+          }));
+
+          osData.onSet(
+            (prev) => ({
+              ...prev,
+              ...localStorageCol[tab],
+            }),
+            true,
+          );
+        }}
+      >
+        <div className="max-w-screen-xl mx-auto px-4 lg:px-8">
+          <TabsList>
+            <TabsTrigger value="spas">SPAs</TabsTrigger>
+            <TabsTrigger value="waivers">Waivers</TabsTrigger>
+          </TabsList>
+        </div>
         <FilterDrawerProvider>
-          <div className="flex flex-col w-full self-center mx-auto max-w-screen-xl xs:flex-row justify-between p-4 lg:px-8">
-            <h1 className="text-xl font-bold mb-4 md:mb-0">Dashboard</h1>
-            {isStateUser({ ...userObj.user, role: userDetails.role }) && (
-              <Link
-                to="/new-submission"
-                className="flex items-center text-white font-bold bg-primary border-none px-10 py-2 rounded cursor-pointer"
-              >
-                <span data-testid="new-sub-button" className="mr-2">
-                  New Submission
-                </span>
-                <PlusIcon className="w-4 h-4" />
-              </Link>
-            )}
-          </div>
-
-          <Tabs
-            value={osData.state.tab}
-            onValueChange={(tab) => {
-              setLocalStorageCol((prev) => ({
-                ...prev,
-                [osData.state.tab]: osData.state,
-              }));
-
-              osData.onSet(
-                (prev) => ({
-                  ...prev,
-                  ...localStorageCol[tab],
-                }),
-                true,
-              );
+          <OsProvider
+            value={{
+              data: osData.data,
+              error: osData.error,
+              isLoading: osData.isLoading,
             }}
           >
-            <div className="max-w-screen-xl mx-auto px-4 lg:px-8">
-              <TabsList>
-                <TabsTrigger value="spas">SPAs</TabsTrigger>
-                <TabsTrigger value="waivers">Waivers</TabsTrigger>
-              </TabsList>
-            </div>
             <TabsContent value="spas">
-              <SpasList />
+              <SpasList oneMacUser={oneMacUser} />
             </TabsContent>
             <TabsContent value="waivers">
-              <WaiversList />
+              <WaiversList oneMacUser={oneMacUser} />
             </TabsContent>
-          </Tabs>
+          </OsProvider>
         </FilterDrawerProvider>
-      </div>
-    </OsProvider>
+      </Tabs>
+    </>
   );
 };
